# Conflicts:
#	Data/Line.dat
#	DefaultWindow/DefaultWindow.vcxproj
#	DefaultWindow/DefaultWindow.vcxproj.filters
#	DefaultWindow/Item.cpp
#	DefaultWindow/Item.h
#	DefaultWindow/LineMgr.cpp
#	DefaultWindow/LineMgr.h
#	DefaultWindow/MainGame.cpp
#	DefaultWindow/ObjMgr.cpp
#	DefaultWindow/ObjMgr.h
#	DefaultWindow/Player.cpp
#	DefaultWindow/Player.h
#	DefaultWindow/Player_Stage1.cpp
#	DefaultWindow/Player_Stage1.h
#	DefaultWindow/Player_Stage3.cpp
#	DefaultWindow/Player_Stage3.h
#	DefaultWindow/Stage1.cpp
#	DefaultWindow/Stage1.h
#	DefaultWindow/Stage3.cpp
#	DefaultWindow/Stage3.h
#	DefaultWindow/stdafx.h
#	Edit/LineMgr.cpp
#	Header/Define.h

+++ DefaultWindow/CollisionMgr.cpp
@@ -34,7 +34,7 @@ bool CCollisionMgr::Check_Sphere(CObj * pDst, CObj * pSrc)
 	float	fWidth  = abs(pDst->Get_Info().fX - pSrc->Get_Info().fX);
 	float	fHeight = abs(pDst->Get_Info().fY - pSrc->Get_Info().fY);
 
-	float	fDiagonal = sqrt(fWidth * fWidth + fHeight * fHeight);
+	float	fDiagonal = (float)sqrt(fWidth * fWidth + fHeight * fHeight);
 
 	float	fRadius = (pDst->Get_Info().fCX + pSrc->Get_Info().fCX) * 0.5f;
